+++ pkg/deployment/features/names.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2020-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2020 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,25 +20,20 @@
 // Author Adam Janikowski
 //
 
-package reconcile
-
-import (
-	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
-	"github.com/rs/zerolog"
-)
+package features
 
 func init() {
-	registerAction(api.ActionTypeMemberRIDUpdate, newMemberRIDUpdate)
+	registerFeature(podNames)
 }
 
-func newMemberRIDUpdate(log zerolog.Logger, action api.Action, actionCtx ActionContext) Action {
-	a := &memberRIDUpdateAction{}
-
-	a.actionImpl = newActionImplDefRef(log, action, actionCtx, defaultTimeout)
-
-	return a
+var podNames = &feature{
+	name:               "short-pod-names",
+	description:        "Enable Short Pod Names",
+	version:            "3.5.0",
+	enterpriseRequired: false,
+	enabledByDefault:   false,
 }
 
-type memberRIDUpdateAction struct {
-	actionEmpty
+func PodNames() Feature {
+	return podNames
 }